+++ mountpoint-s3-fs/src/fs/error.rs
@@ -168,6 +168,7 @@ impl ToErrno for InodeError {
             // Not obvious what InodeNotWritable, InodeAlreadyWriting, InodeNotReadableWhileWriting should be.
             // EINVAL or EROFS would also be reasonable -- but we'll treat them like sealed files.
             InodeError::InodeNotWritable(_) => libc::EPERM,
+            InodeError::ReadOnlyMount() => libc::EROFS,
             InodeError::InodeInvalidWriteStatus(_) => libc::EPERM,
             InodeError::InodeAlreadyWriting(_) => libc::EPERM,
             InodeError::InodeNotReadableWhileWriting(_) => libc::EPERM,