+++ mountpoint-s3-fs/src/fs/error_metadata.rs
@@ -22,5 +22,6 @@ pub struct ErrorMetadata {
 /// NOTE: the event log API is not stable and is subject to breaking changes
 pub const MOUNTPOINT_ERROR_CLIENT: &str = "error.client";
 pub const MOUNTPOINT_ERROR_LOOKUP_NONEXISTENT: &str = "error.fs.lookup_nonexistent";
+pub const MOUNTPOINT_ERROR_FS_READ_ONLY: &str = "error.fs.read_only";
 pub const MOUNTPOINT_ERROR_INTERNAL: &str = "error.internal";
 pub const MOUNTPOINT_EVENT_READY: &str = "event.ready";